+++ src/bsvar.cpp
@@ -23,6 +23,11 @@ Rcpp::List bsvar_cpp(
   const bool        show_progress = true
 ) {
 
+  std::string oo = "";
+  if ( thin != 1 ) {
+    oo      = ordinal(thin) + " ";
+  }
+  
   // Progress bar setup
   vec prog_rep_points = arma::round(arma::linspace(0, S, 50));
   if (show_progress) {
@@ -32,7 +37,7 @@ Rcpp::List bsvar_cpp(
     Rcout << " Gibbs sampler for the SVAR model                 |" << endl;
     Rcout << "**************************************************|" << endl;
     Rcout << " Progress of the MCMC simulation for " << S << " draws" << endl;
-    Rcout << "    Every " << thin << "th draw is saved via MCMC thinning" << endl;
+    Rcout << "    Every " << oo << "draw is saved via MCMC thinning" << endl;
     Rcout << " Press Esc to interrupt the computations" << endl;
     Rcout << "**************************************************|" << endl;
   }